[merge] midend

+++ include/ir/utils_ir.hpp
@@ -28,7 +28,7 @@ inline std::enable_if_t<std::is_base_of_v<Value, T>, bool> isa(const Value* valu
 
 
 //! get machine code for float
-inline std::string getMC(double f){
+inline std::string getMC(float f){
     double d = f;
     unsigned long mrf = *reinterpret_cast<unsigned long*>(&d);
     std::stringstream ss;
@@ -36,4 +36,12 @@ inline std::string getMC(double f){
     std::string res = "0x" + ss.str();
     return res;
 }
+// inline std::string getMC(double f){
+//     double d = f;
+//     unsigned long mrf = *reinterpret_cast<unsigned long*>(&d);
+//     std::stringstream ss;
+//     ss << std::hex << std::uppercase << std::setfill('0') << std::setw(16) << mrf;
+//     std::string res = "0x" + ss.str();
+//     return res;
+// }
 }  // namespace ir
